Completed editor. Changed ParticleSystem class to perform better with editor. Added Fragment class as a base class for terrain, entity, animal, particlesystem, seaeffect. Added some support for filesystem management.

+++ fruithunter/Source/Utility/Development/TerrainBatch.cpp
@@ -95,6 +95,6 @@ void TerrainBatch::draw_onlyMesh() {
 		m_terrains[i]->draw_onlyMesh();
 }
 
-float3 TerrainBatch::getSpawnpoint(int terrainIndex) {
+float3 TerrainBatch::getSpawnpoint(size_t terrainIndex) {
 	return m_terrains[terrainIndex]->getRandomSpawnPoint();
 }
